requirements.txt
python-telegram-bot>=20.0
python-dotenv
SQLAlchemy
psycopg2-binary
httpx
beautifulsoup4
lxml
playwright
aioredis
alembic
aiokafka
